% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_osm.R
\name{create_terrain_osm}
\alias{create_terrain_osm}
\title{Create 3D terrain with OpenStreamMap data overlaying.}
\usage{
create_terrain_osm(
  center_ll = c(102.929, 30.15),
  half_width = 0.05,
  elevat_z = 12,
  add_highway = TRUE,
  add_waterway = TRUE,
  add_buildings = FALSE,
  add_parkings = FALSE,
  add_tourisms = FALSE,
  add_places = TRUE,
  plot_3d = TRUE,
  plot_3d_zscale = 20
)
}
\arguments{
\item{center_ll, }{Center longitude and latitude, c(lon, lat).}

\item{half_width, }{Half width of map region in degrees.}

\item{elevat_z, }{The zoom level to return. The zoom ranges from 1 to 14.
Resolution of the resultant raster is determined by
the zoom and latitude.}

\item{add_highway, }{add highway lines.}

\item{add_waterway, }{add waterway lines.}

\item{add_buildings, }{add building polygons, not recommend for large city.}

\item{add_parkings, }{add parking polygons.}

\item{add_tourisms, }{add tourism polygons.}

\item{add_places, }{add place labels.}

\item{plot_3d, }{plot 3d map, if not, plot 2d map.}

\item{plot_3d_zscale, }{Default '1'. The ratio between the x and y spacing
(which are assumed to be equal) and the z axis.
For example, if the elevation levels are in units
of 1 meter and the grid values are separated by
10 meters, 'zscale' would be 10. Adjust the zscale
down to exaggerate elevation features.}
}
\value{
rayshader graphics.
}
\description{
Create 3D terrain with OpenStreamMap data overlaying.
}
\examples{
create_terrain_osm()
}
